Re-did ammo counter
currently has icons on the spritesheet for a magnum, rifle, and 9mm bullet, shotgun shell, arrowhead, and a placeholder "thow" icon. That will need to be updated as we add weapons to the game. For now, I have the ammo labeled with a standard label (that hasn't been formatted) but we may want to use a RichTextLabel in the future. Didn't do anything for the enemy's ammo counter yet. Deleted unused nodes.