Added sandy tileset
I tries to make the sky look smoggy and make the sand not quite flat. I initially tried to make a human skull wal for a psycho raider gang area, but that didn't work. Then I tried to make spikes, because spikes are mean and scary, which had similar results. The monstrosities you see on the side and top of the screen are supposed to be rusty iron girders, but those came out pretty rough too.