Brainstorming
I'm not sure how we want to do cars and boats, but if we use them as a sort of mobile base where the players can fast travel, sleep, and store stuff, I've fleshed out a little bit of what cars might look like and what traits they would have. I also added the raccon's treasure location to the locations list, as well as a few other new ideas.