Started adding text to segments, pause menu can be escaped with escape button
I had to check if the pause button was just released, but it works normally unless the player holds the pause button (which I actually kind of like and we can call it a quick pause). I also started to work on putting the inventory into the segments, but that is still a work in progress.