Added assets for inventory options popup
It will effectively be a miniature version of the inventory screen. Pressing left (for p1) or right (for p2), or clicking back on the inventory screen itself should close it, and pressing enter or clicking should do one of its actions. Its text needs to be different based on the type of item and the situation (certain items may have the option to "equip", but if they're equipped they need to say "unequip", and maybe items can be repaired but not given to a teammate at a workbench).