Created basic bot, other stuff
I put the aim and move bots together so we can see how they work together as a basic AI. It would stop shooting after 2 shots, so I added a couple of lines to try and add a "desperation mode", which is less than exceptional. I also finally changed the enemy ammo counter to actually reflect how many shots the enemy has left. I also added some items to the openworld to try and debug the inventory screen and to hopefully work on scrolling through a full inventory.