Imported Addon for Dialog Trees
It seems like a pretty useful way to build dialog trees, which we aren't quite dealing with yet. Trying to set it to "active" in the project settings leads to a problem where Godot says "Base type is not EditorPlugin"
It's not a priority yet, but I think it will be good once we get it up and running and make quest creation much easier.
It's under the MIT license, so bringing it in is a non-issue.
Link to the GitHub:
https://github.com/StraToN/godot-dialogtreetools/wiki/How-to-use