Added Item Description to Inventory Screen ,pogress towards adding Icon as well
The item description is now perfectly functional. I added item description and image path to the item class to make this easier and updated the unit tests to keep them from breaking the game (though they are not  testing the new changes). I also changed the sprite to make the inventory screen look a little better. I am very close to getting the inventory screen to display the icon, which means that after that all that will be left will be connecting signals and implementing scrolling before the inventory screen is complete!