Rotated hockey scene
Man, look at all that room for activities. I left a bit of room on the edge to animate different arenas and enough on top for a HUD- I think we need the score, the time remaining, the period, and the status of the players on each bench (if they can come into the game or not). I'll get to work on that HUD too.
Closed #180