Early stages of tag team mechanic
Players can now switch between player 1 and player 2. Still need to make the players have their own health, ammo, and stamina, and I need to add the countdown for the number of switches players can do. We may also want to add a short cooldown where the player is invincible and can't switch.