Astrted working on different way of solving the musicMinigame problem
Spawning new notes wasn't working properly, so I started workig on a system where we have a carousel of the same notes coming down and recycling. It breaks now so everything is commented out at the moment, and I'm not sure if this is the right way to go about it so I've kept all of the spawning code as well.